修改： Generator部分代码中目录结构，使之支持从根目录运行 新增： 根目录下一键运行转换为字体脚本，已测试； 根目录下一键生成字体脚本，未测试。
