+++ Generator/src/bitmap.py
@@ -2,6 +2,7 @@ import os
 from PIL import Image
 
 def bitmap_to_svg(bitmap_path, svg_path):
+    # print(bitmap_path, svg_path)
     # 转换图像格式为 BMP
     bmp_path = bitmap_path.replace('.png', '.bmp')
     image = Image.open(bitmap_path)
@@ -14,4 +15,9 @@ def bitmap_to_svg(bitmap_path, svg_path):
     # 作用：控制输出矢量图像中曲线角的平滑度。数值越高，曲线越平滑（越少棱角）。
     # 参数：最大角度设置，可达最高4.0。
     # --opttolerance 0.4 --alphamax 2.5
-    os.system(f'.\\potrace\\bin\\potrace.exe {bmp_path} -s -o {svg_path}')
+
+    # Windows版本：直接运行下述代码
+    os.system(f'.\\Generator\\potrace\\bin\\potrace.exe {bmp_path} -s -o {svg_path}')
+
+    # Linux版本：需要安装potrace（见Generator README），并注册进入路径
+    # os.system(f'potrace {bmp_path} -s -o {svg_path}')
